add EmintBackend, migrate eth_ methods to EmintBackend
